answer selection add up to scores correctly with level navigation

+++ app/src/main/java/prog/com/quizapp/utils/Constants.java
@@ -6,8 +6,8 @@ package prog.com.quizapp.utils;
 
 public class Constants {
 
-    public static final double COMPLETE_MARKS = 10;
-    public static final double MARK_FOR_ONE_QUESTION = 1;
+    public static final int COMPLETE_MARKS = 10;
+    public static final int MARK_FOR_ONE_QUESTION = 1;
     public static final int QUESTIONS_FOR_LEVEL = 10;
     public static final int LEVELS = 3;
     public static final int LEVEL_ONE_PASS_MARK = 5;
